Enhance LendingMarketAggregator with user obligations and protocol integration
- Refactor LendingMarketAggregator to support generic client type
- Implement get_user_obligations method to fetch obligations across protocols
- Add print_obligations method for formatted obligation display
- Update type conversions for market price and obligation types
- Modify main.rs to demonstrate new obligation fetching functionality

+++ drift/src/client.rs
@@ -123,7 +123,7 @@ impl<C: Clone + Deref<Target = impl Signer>> DriftClient<C> {
                         .to_string();
                     (
                         name.clone(),
-                        market.historical_oracle_data.last_oracle_price,
+                        market.historical_oracle_data.last_oracle_price as u64,
                         market.mint,
                         market.decimals,
                         name,
